Developing dressed types; need to set visibility to 'hidden'. Is that a problem for MacOS?

+++ tests/test_PR028_add_dressed_types.py
@@ -19,7 +19,7 @@ def test_highlevel():
     assert repr(a) == "<Array [[1.1 2.2 3.3] [] ... [7.7 8.8 9.9]] type='5 * var * float64'>"
     assert str(a) == "[[1.1 2.2 3.3] [] [4.4 5.5] [6.6] [7.7 8.8 9.9]]"
 
-    b = awkward1.Array(numpy.arange(100))
+    b = awkward1.Array(numpy.arange(100, dtype=numpy.int64))
     assert repr(b) == "<Array [0 1 2 3 4 5 6 ... 93 94 95 96 97 98 99] type='100 * int64'>"
     assert str(b) == "[0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 ... 85 86 87 88 89 90 91 92 93 94 95 96 97 98 99]"
